Another failure
I don't know if its a bug or what, but even after I got the depth part working (I forgot to inverse faces of the quad before) I can't use it as a mask because the it still somehow effects the lighting of object near it, regardless of the render options I give it to disable that. I might have to give up on this for now and move onto a different thing.