Fix 1-wide rotation properly
When the DoublePiece is in a 1-wide shaft, only vertical orientations
are allowed. However, it should still be possible to flip the piece 180
degrees. Players expect to be able to do this the same way they normally
do; by double-tapping a rotation button. To achieve this effect without
passing through a horizontal orientation, we keep track of rotation
state within the DoublePiece.